+++ spec/factories/fields.rb
@@ -3,5 +3,11 @@ FactoryGirl.define do
     name 'Name'
     liquor_name 'name'
     type Kms::StringField.name
+    factory :has_many_field do
+      type Kms::HasManyField.name
+      name 'Comments'
+      liquor_name 'comments'
+      class_name { FactoryGirl.create(:associated_model).id }
+    end
   end
 end